[compiler] Fix compiler error by static_casting to quint32 and renaming type
This includes a static_cast to quint32 for sizeof and correction to header to use
size_t instead of quint32.

+++ src/engine/contactstransientstore.h
@@ -52,7 +52,7 @@ public:
         friend class ContactsTransientStore;
 
     protected:
-        const_iterator(const MemoryTable *table, quint32 position);
+        const_iterator(const MemoryTable *table, size_t position);
 
     public:
         const_iterator(const const_iterator &other);
